feat(core): extract tests to their own file
